requirements.txt
Flask
Flask-APScheduler
pexpect
PyYAML
Flask-Bootstrap
Flask-Moment
Flask-SQLAlchemy
flask-wtf
pymysql
requests
Flask-MySQL
Flask-Login
flask-assets
tornado
paramiko
pyecharts
flask-jsglue